requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.32.4
beautifulsoup4==4.13.4
openai==1.3.7
pydantic==2.11.7
python-dotenv==1.0.0
httpx==0.28.1
playwright==1.40.0
asyncio
pandas==2.1.3
openpyxl==3.1.2 